refactor: move sort and filter to separate components
